+++ src/state/activeBoards/activeBoardCalculationSaga.test.ts
@@ -1,10 +1,10 @@
 import { expectSaga } from 'redux-saga-test-plan';
 import { select } from 'redux-saga/effects';
-import { getBoards } from '../../selectors/AppStateSelectors';
-import { CALCULATE_ALLOWED_BOARDS, SET_ALLOWED_BOARDS } from '../../activeBoards/activeBoardsActions';
-import activeBoardsCalculationSaga from '../activeBoardsCalculationSaga';
-import unfinishedBoardMock from '../../../__mocks__/unfinishedBoardMock';
-import { circleFinishedBoardMock } from '../../../__mocks__/finishedBoardMock';
+import { getBoards } from '../selectors/AppStateSelectors';
+import { CALCULATE_ALLOWED_BOARDS, SET_ALLOWED_BOARDS } from './activeBoardsActions';
+import activeBoardsCalculationSaga from './activeBoardsCalculationSaga';
+import unfinishedBoardMock from '../../__mocks__/unfinishedBoardMock';
+import { circleFinishedBoardMock } from '../../__mocks__/finishedBoardMock';
 
 describe( 'activeBoardCalculationSaga', () => {
     it( 'should dispatch setAllowedBoard Action, normal case', () => {
