+++ src/state/game/checkGameFinishedSaga.test.ts
@@ -1,13 +1,13 @@
 import { expectSaga } from 'redux-saga-test-plan';
-import checkGameFinishedSaga from '../checkGameFinishedSaga';
-import { getBoards } from '../../selectors/AppStateSelectors';
+import checkGameFinishedSaga from './checkGameFinishedSaga';
+import { getBoards } from '../selectors/AppStateSelectors';
 import { select } from 'redux-saga/effects';
-import { CHECK_GAME_FINISHED, GAME_FINISHED } from '../../game/gameAction';
-import { Player } from '../../AppState';
-import { SET_ALLOWED_BOARDS } from '../../activeBoards/activeBoardsActions';
-import { circleFinishedBoardMock, crossFinishedBoardMock } from '../../../__mocks__/finishedBoardMock';
-import unfinishedBoardMock from '../../../__mocks__/unfinishedBoardMock';
-import { SAVE_GAME_DATA } from '../saveFinishedGameDataActions';
+import { CHECK_GAME_FINISHED, GAME_FINISHED } from './gameAction';
+import { Player } from '../AppState';
+import { SET_ALLOWED_BOARDS } from '../activeBoards/activeBoardsActions';
+import { circleFinishedBoardMock, crossFinishedBoardMock } from '../../__mocks__/finishedBoardMock';
+import unfinishedBoardMock from '../../__mocks__/unfinishedBoardMock';
+import { SAVE_GAME_DATA } from '../finishedGames/saveFinishedGameDataActions';
 
 describe( 'checkGameFinishedSaga', () => {
     it( 'should dispatch gameFinished action with Circle and' +
